- реализации business логики part 3

+++ src/main/java/ru/nemodev/wifi/analyzer/service/wifi/report/WifiAnalyzeReportServiceImpl.java
@@ -1,16 +1,18 @@
-package ru.nemodev.wifi.analyzer.service.wifi_report;
+package ru.nemodev.wifi.analyzer.service.wifi.report;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import ru.nemodev.wifi.analyzer.entity.wifi_report.WifiAnalyzeReport;
-import ru.nemodev.wifi.analyzer.repository.wifi_report.WifiAnalyzeReportRepository;
+import ru.nemodev.wifi.analyzer.entity.wifi.report.WifiAnalyzeReport;
+import ru.nemodev.wifi.analyzer.repository.wifi.report.WifiAnalyzeReportRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-
+/**
+ * created by simanov-an
+ */
 public class WifiAnalyzeReportServiceImpl implements WifiAnalyzeReportService {
 
     private final WifiAnalyzeReportRepository wifiAnalyzeReportRepository;
@@ -22,8 +24,15 @@ public class WifiAnalyzeReportServiceImpl implements WifiAnalyzeReportService {
     @Override
     @Transactional
     public WifiAnalyzeReport create(WifiAnalyzeReport wifiAnalyzeReport) {
-        wifiAnalyzeReport.setCreationDate(LocalDateTime.now());
-        return wifiAnalyzeReportRepository.saveAndFlush(wifiAnalyzeReport);
+        final LocalDateTime nowDate = LocalDateTime.now();
+        wifiAnalyzeReport.setCreationDate(nowDate);
+        wifiAnalyzeReport.getWifiAnalyzeInfoList()
+                .forEach(wifiAnalyzeInfo -> {
+                    wifiAnalyzeInfo.setCreationDate(nowDate);
+                    wifiAnalyzeInfo.setWifiAnalyzeReport(wifiAnalyzeReport);
+                });
+
+        return wifiAnalyzeReportRepository.save(wifiAnalyzeReport);
     }
 
     @Override
@@ -35,9 +44,12 @@ public class WifiAnalyzeReportServiceImpl implements WifiAnalyzeReportService {
     @Override
     @Transactional
     public void deleteById(String id) {
-        wifiAnalyzeReportRepository.findById(id).orElseThrow(()->
-                new IllegalArgumentException(String.format("Не удалось найти wifi-отчет с ID=%s!", id)));
-        wifiAnalyzeReportRepository.deleteById(id);
+        if (wifiAnalyzeReportRepository.existsById(id)) {
+            wifiAnalyzeReportRepository.deleteById(id);
+        }
+        else {
+            throw new IllegalArgumentException(String.format("Не удалось найти wifi-report с ID=%s!", id));
+        }
     }
 
     @Override
